Структура проекта todo
