+++ src/components/App/App.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import InputItem from '../InputItem/InputItem';
+import ItemList from '../ItemList/ItemList';
+import Footer from '../Footer/Footer';
+
+const App = () => (<div>
+  <h2>todos</h2>
+  <InputItem />
+  <ItemList />
+  <Footer />
+</div>);
+
+export default App;
