+++ src/components/InputItem/InputItem.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const InputItem = () => (
+<div>
+  <p>Введи новое дело</p>
+</div>);
+
+export default InputItem;
