+++ src/components/Item/Item.js
@@ -1,0 +1,14 @@
+import React from 'react';
+const Item = () => (
+  <ol>
+    <li>Сделать зарядку</li>
+    <li>Почистить зубы</li>
+    <li>Съесть завтрак</li>
+    <li>Пойти на работу</li>
+    <li>Забрать пальто из химчистки</li>
+    <li>Отвезти кота в вет клинику</li>
+    <li>Забрать посылку с почты</li>
+    <li>Купить хлеб</li>
+  </ol>
+);
+export default Item;
