+++ src/components/ItemList/ItemList.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Item from '../Item/Item';
+const ItemList = () => (
+<div>
+  <Item />
+</div>
+);
+export default ItemList;
